+++ src/components/RegistrationModal.tsx
@@ -1,13 +1,10 @@
 import React from "react";
-import { Alert, Button, Form, Input, Popover } from "antd";
-import { useAppSelector } from "../redux/hooks";
-import UserAvatar from "./UserAvatar";
-import { core } from "@dsnp/sdk";
-import * as dsnp from "../services/dsnp";
+import { Button, Popover } from "antd";
 import * as registry from "@dsnp/sdk/core/contracts/registry";
-import { HexString, Profile } from "../utilities/types";
-import { Registration } from "@dsnp/sdk/core/contracts/registry";
+import { HexString } from "../utilities/types";
 import { DSNPUserURI } from "@dsnp/sdk/dist/types/core/identifiers";
+import EditRegistrationAccordion from "./EditRegistrationAccordion";
+import Logout from "./Logout";
 
 interface RegistrationModalProps {
   children: JSX.Element;
@@ -15,7 +12,7 @@ interface RegistrationModalProps {
   registrations: registry.Registration[];
   walletAddress: HexString;
   onIdResolved: (uri: DSNPUserURI) => void;
-  onCancel: () => void;
+  logout: () => void;
 }
 
 const RegistrationModal = ({
@@ -24,7 +21,7 @@ const RegistrationModal = ({
   registrations,
   walletAddress,
   onIdResolved,
-  onCancel,
+  logout,
 }: RegistrationModalProps): JSX.Element => {
   const [hasRegistrations, setHasRegistrations] = React.useState<boolean>(
     false
@@ -34,18 +31,6 @@ const RegistrationModal = ({
     isCreatingRegistration,
     setIsCreatingRegistration,
   ] = React.useState<boolean>(true);
-
-  const [registrationError, setRegistrationError] = React.useState<
-    string | undefined
-  >(undefined);
-
-  const [selectedRegistration, setRegistration] = React.useState<
-    registry.Registration | undefined
-  >(undefined);
-
-  const profiles: Record<HexString, Profile> = useAppSelector(
-    (state) => state.profiles?.profiles || {}
-  );
 
   if (registrations.length) {
     // Flip to select registration if we see registrations increase
@@ -61,136 +46,6 @@ const RegistrationModal = ({
     }
   }
 
-  // convert registration to uri before calling callback
-  const commitRegistration = () => {
-    if (!selectedRegistration) return;
-    onIdResolved(selectedRegistration.dsnpUserURI);
-  };
-
-  // create new DSNP registration
-  const register = async (formValues: { handle: string }) => {
-    try {
-      const userURI = await dsnp.createNewDSNPRegistration(
-        walletAddress,
-        formValues.handle
-      );
-      onIdResolved(core.identifiers.convertToDSNPUserId(userURI).toString());
-    } catch (error) {
-      console.error(error);
-      setRegistrationError(
-        `Error registering: ${error.message || error.toString()}`
-      );
-    }
-  };
-
-  // Find the profile photo a given DSNP registration
-  const iconForRegistration = (
-    registration: registry.Registration
-  ): string | undefined =>
-    profiles[
-      core.identifiers.convertToDSNPUserId(registration.dsnpUserURI).toString()
-    ]?.icon?.[0].href;
-
-  /**
-   * New Registration Form
-   */
-  const registerNewHandleForm = (
-    <Form
-      name="createHandle"
-      className="RegistrationModal__createHandle"
-      initialValues={{
-        remember: true,
-      }}
-      onFinish={register}
-    >
-      {registrationError && (
-        <>
-          <Alert message={registrationError} type="error" />
-        </>
-      )}
-      <p>Please create a handle:</p>
-      <Form.Item
-        name="handle"
-        className="RegistrationModal__handleInput"
-        rules={[
-          {
-            required: true,
-            message: "Handle cannot be blank",
-          },
-        ]}
-      >
-        <Input />
-      </Form.Item>
-      <div className="RegistrationModal__footer">
-        {hasRegistrations && (
-          <button
-            onClick={(e) => {
-              e.preventDefault();
-              setIsCreatingRegistration(false);
-            }}
-            className="RegisrationModal__changeMode"
-          >
-            Select Existing Handle
-          </button>
-        )}
-        <Button
-          type="primary"
-          htmlType="submit"
-          className="RegistrationModal__footerBtn"
-        >
-          Choose handle
-        </Button>
-      </div>
-    </Form>
-  );
-
-  /**
-   * Select Existing Handle UI
-   */
-  const selectHandleContent = (
-    <div>
-      <p>You have multiple handles associated with your account.</p>
-      <h3>Select an account:</h3>
-      <div className="RegistrationModal__registrations">
-        {registrations.map((registration: Registration) => (
-          <button
-            className={`RegistrationModal__registration${
-              registration === selectedRegistration
-                ? " RegistrationModal__registration--selected"
-                : ""
-            }`}
-            onClick={() => setRegistration(registration)}
-            key={registration.dsnpUserURI}
-          >
-            <UserAvatar
-              icon={iconForRegistration(registration)}
-              profileAddress={registration.dsnpUserURI}
-              avatarSize="small"
-            />
-            {`@${registration.handle}`}
-          </button>
-        ))}
-      </div>
-      <div className="RegistrationModal__footer">
-        <button
-          onClick={() => setIsCreatingRegistration(true)}
-          className="RegisrationModal__changeMode"
-        >
-          Create New Handle
-        </button>
-        <Button
-          className="RegistrationModal__footerBtn"
-          key="post"
-          type="primary"
-          disabled={!selectedRegistration}
-          onClick={commitRegistration}
-        >
-          Select Handle
-        </Button>
-      </div>
-    </div>
-  );
-
   /**
    * Create a popover with either registration or handle selection UI.
    */
@@ -203,12 +58,20 @@ const RegistrationModal = ({
           <Button
             type="link"
             className="RegistrationModal__cancel"
-            onClick={onCancel}
+            onClick={logout}
           >
             Cancel
           </Button>
-          <h2>Welcome!</h2>
-          {isCreatingRegistration ? registerNewHandleForm : selectHandleContent}
+          <p>Successfully connected to Wallet</p>
+          <EditRegistrationAccordion
+            isCreatingRegistration={isCreatingRegistration}
+            walletAddress={walletAddress}
+            onIdResolved={onIdResolved}
+            hasRegistrations={hasRegistrations}
+            setIsCreatingRegistration={setIsCreatingRegistration}
+            registrations={registrations}
+          />
+          <Logout logout={logout} />
         </div>
       }
     >
